pyproject.toml
[tool.poetry]
name = "b2broker"
version = "0.1.0"
description = ""
authors = ["d.rzhevskii <d.rzhevskii@prequelapp.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
django-filter = "^24.2"
mysql-connector-python = "^9.0.0"
djangorestframework-jsonapi = "^7.0.2"
drf-spectacular = "^0.27.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
pytest-cov = "^5.0.0"
flake8 = "^7.1.0"
black = "^24.4.2"
isort = "^5.13.2"
pre-commit = "^3.7.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

